wire signin + move signout to header
